+++ streams-contrib/streams-provider-facebook/src/main/java/org/apache/streams/facebook/feed/FacebookPublicFeedXmlActivitySerializer.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package com.facebook.feed;
+package org.apache.streams.facebook.feed;
 
 /**
  * Created with IntelliJ IDEA.
