+++ streams-contrib/streams-provider-facebook/src/main/jsonschema/org/apache/streams/facebook/graph/Post.json
