+++ streams-contrib/streams-provider-gnip/gnip-edc-reddit/src/test/resources/redditTest.xml
@@ -1,1 +1,1 @@
-<entry xmlns="http://www.w3.org/2005/Atom" xmlns:activity="http://activitystrea.ms/spec/1.0/"		xmlns:service="http://activitystrea.ms/service-provider" xmlns:thr="http://purl.org/syndication/thread/1.0"		xmlns:gnip="http://www.gnip.com/schemas/2010">		<id>1088dt</id>		<published>2012-09-21T03:28:15Z</published>		<updated>2012-09-21T03:28:15Z</updated>		<title>Precious_Zest posted an article Reddit</title>		<category term="ArticlePosted" label="Article Posted"/>		<link rel="alternate" type="text/html" href="http://www.reddit.com/r/atheism/comments/1088dt/a_classy_new_way_to_respond_to_stupid_fundie/"/>		<source>			<link rel="self" type="application/json" href="http://com.facebook.api.reddit.com/search.json?q=%22zest%22%20%20%22%22%20%20--%22%22&amp;limit=100&amp;after=t3_1brddb"/>			<title>Reddit - Keyword - Search (Recent) - "zest"  ""  --""</title>			<updated>2013-08-28T17-54-04Z</updated>			<gnip:rule xmlns:gnip="http://www.gnip.com/schemas/2010">"zest"  ""  --""            </gnip:rule>		</source>		<service:provider>			<name>Reddit</name>			<uri>www.reddit.com</uri>		</service:provider>		<activity:verb>http://activitystrea.ms/schema/1.0/post</activity:verb>		<activity:object>			<activity:object-type>http://activitystrea.ms/schema/1.0/article</activity:object-type>			<id>1088dt</id>			<title>A classy new way to respond to stupid fundie claims (X-post from r/funny) </title>			<link rel="alternate" type="text/html" href="http://www.reddit.com/r/atheism/comments/1088dt/a_classy_new_way_to_respond_to_stupid_fundie/"/>			<link rel="enclosure" type="text/html" href="http://i.imgur.com/ZphAO.png"/>			<link rel="preview" type="image/jpeg" href="http://a.thumbs.redditmedia.com/iL3sa_rwh2j-WA7H.jpg"/>			<gnip:statistics upVotes="220" downVotes="69"/>		</activity:object>		<author>			<name>Precious_Zest</name>			<uri>http://www.reddit.com/user/Precious_Zest</uri>		</author>		<activity:actor>			<activity:object-type>http://activitystrea.ms/schema/1.0/person</activity:object-type>			<link rel="alternate" type="text/html" length="0" href="http://www.reddit.com/user/Precious_Zest"/>			<id>http://www.reddit.com/user/Precious_Zest</id>		</activity:actor>		<gnip:matching_rules>			<gnip:matching_rule rel="source" tag="Zest">"zest"  ""  --""</gnip:matching_rule>		</gnip:matching_rules>	</entry>
+<entry xmlns="http://www.w3.org/2005/Atom" xmlns:activity="http://activitystrea.ms/spec/1.0/"		xmlns:service="http://activitystrea.ms/service-provider" xmlns:thr="http://purl.org/syndication/thread/1.0"		xmlns:gnip="http://www.gnip.com/schemas/2010">		<id>1088dt</id>		<published>2012-09-21T03:28:15Z</published>		<updated>2012-09-21T03:28:15Z</updated>		<title>Precious_Zest posted an article Reddit</title>		<category term="ArticlePosted" label="Article Posted"/>		<link rel="alternate" type="text/html" href="http://www.reddit.com/r/atheism/comments/1088dt/a_classy_new_way_to_respond_to_stupid_fundie/"/>		<source>			<link rel="self" type="application/json" href="http://org.apache.streams.facebook.api.reddit.com/search.json?q=%22zest%22%20%20%22%22%20%20--%22%22&amp;limit=100&amp;after=t3_1brddb"/>			<title>Reddit - Keyword - Search (Recent) - "zest"  ""  --""</title>			<updated>2013-08-28T17-54-04Z</updated>			<gnip:rule xmlns:gnip="http://www.gnip.com/schemas/2010">"zest"  ""  --""            </gnip:rule>		</source>		<service:provider>			<name>Reddit</name>			<uri>www.reddit.com</uri>		</service:provider>		<activity:verb>http://activitystrea.ms/schema/1.0/post</activity:verb>		<activity:object>			<activity:object-type>http://activitystrea.ms/schema/1.0/article</activity:object-type>			<id>1088dt</id>			<title>A classy new way to respond to stupid fundie claims (X-post from r/funny) </title>			<link rel="alternate" type="text/html" href="http://www.reddit.com/r/atheism/comments/1088dt/a_classy_new_way_to_respond_to_stupid_fundie/"/>			<link rel="enclosure" type="text/html" href="http://i.imgur.com/ZphAO.png"/>			<link rel="preview" type="image/jpeg" href="http://a.thumbs.redditmedia.com/iL3sa_rwh2j-WA7H.jpg"/>			<gnip:statistics upVotes="220" downVotes="69"/>		</activity:object>		<author>			<name>Precious_Zest</name>			<uri>http://www.reddit.com/user/Precious_Zest</uri>		</author>		<activity:actor>			<activity:object-type>http://activitystrea.ms/schema/1.0/person</activity:object-type>			<link rel="alternate" type="text/html" length="0" href="http://www.reddit.com/user/Precious_Zest"/>			<id>http://www.reddit.com/user/Precious_Zest</id>		</activity:actor>		<gnip:matching_rules>			<gnip:matching_rule rel="source" tag="Zest">"zest"  ""  --""</gnip:matching_rule>		</gnip:matching_rules>	</entry>
